requirements.txt
accelerate==0.24.1
certifi==2023.7.22
charset-normalizer==3.3.2
diffusers==0.21.4
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==2.0.7
zipp==3.17.0
